relocated favorites service
